Change to use int64

+++ Team05/Code05/src/spa/src/models/AST/factor_node/StringExpr.cpp
@@ -1,6 +1,9 @@
 #include "StringExpr.h"
 
-typedef long long ll;
+#include <cstdint>
+#include <vector>
+
+typedef int64_t ll;
 
 const ll kPrime = 195262363;
